requirements.txt
facepy
argparse
inflection
PyExecJS
requests
